+++ elasticsearch/client/snapshot.py
@@ -1,4 +1,4 @@
-from .utils import NamespacedClient, query_params, _make_path
+from .utils import NamespacedClient, query_params, _make_path, SKIP_IN_PATH
 
 class SnapshotClient(NamespacedClient):
     @query_params('master_timeout', 'wait_for_completion')
@@ -15,6 +15,9 @@ class SnapshotClient(NamespacedClient):
         :arg wait_for_completion: Should this request wait until the operation
             has completed before returning, default False    
         """
+        for param in (repository, snapshot):
+            if param in SKIP_IN_PATH:
+                raise ValueError("Empty value passed for a required argument.")
         _, data = self.transport.perform_request('PUT', _make_path('_snapshot',
             repository, snapshot), params=params, body=body)
         return data
@@ -30,6 +33,9 @@ class SnapshotClient(NamespacedClient):
         :arg master_timeout: Explicit operation timeout for connection to master
             node    
         """
+        for param in (repository, snapshot):
+            if param in SKIP_IN_PATH:
+                raise ValueError("Empty value passed for a required argument.")
         _, data = self.transport.perform_request('DELETE',
             _make_path('_snapshot', repository, snapshot), params=params)
         return data
@@ -45,6 +51,9 @@ class SnapshotClient(NamespacedClient):
         :arg master_timeout: Explicit operation timeout for connection to master
             node    
         """
+        for param in (repository, snapshot):
+            if param in SKIP_IN_PATH:
+                raise ValueError("Empty value passed for a required argument.")
         _, data = self.transport.perform_request('GET', _make_path('_snapshot',
             repository, snapshot), params=params)
         return data
@@ -60,6 +69,8 @@ class SnapshotClient(NamespacedClient):
             node
         :arg timeout: Explicit operation timeout    
         """
+        if repository in SKIP_IN_PATH:
+            raise ValueError("Empty value passed for a required argument 'repository'.")
         _, data = self.transport.perform_request('DELETE',
             _make_path('_snapshot', repository), params=params)
         return data
@@ -92,6 +103,9 @@ class SnapshotClient(NamespacedClient):
             node
         :arg timeout: Explicit operation timeout    
         """
+        for param in (repository, body):
+            if param in SKIP_IN_PATH:
+                raise ValueError("Empty value passed for a required argument.")
         _, data = self.transport.perform_request('PUT', _make_path('_snapshot',
             repository), params=params, body=body)
         return data
@@ -110,6 +124,9 @@ class SnapshotClient(NamespacedClient):
         :arg wait_for_completion: Should this request wait until the operation
             has completed before returning, default False    
         """
+        for param in (repository, snapshot):
+            if param in SKIP_IN_PATH:
+                raise ValueError("Empty value passed for a required argument.")
         _, data = self.transport.perform_request('POST', _make_path('_snapshot',
             repository, snapshot, '_restore'), params=params, body=body)
         return data
@@ -143,6 +160,8 @@ class SnapshotClient(NamespacedClient):
             node
         :arg timeout: Explicit operation timeout
         """
+        if repository in SKIP_IN_PATH:
+            raise ValueError("Empty value passed for a required argument 'repository'.")
         _, data = self.transport.perform_request('POST', _make_path('_snapshot',
             repository, '_verify'), params=params)
         return data